requirements-dev.txt
mypy~=0.991
mypy-extensions~=0.4.3
Pillow~=9.3.0
tomli~=2.0.1
types-Pillow~=9.3.0.4
typing_extensions~=4.4.0
